-created user authorization

+++ backend/routes/posts.js
@@ -1,7 +1,8 @@
 const express = require('express');
 const multer = require('multer');
 const Post = require('../models/post');
-const checkAuth = require('../middleware/check-auth')
+const checkAuth = require('../middleware/check-auth');
+const { ResolvedReflectiveFactory } = require('@angular/core');
 
 const router = express.Router();
 
@@ -32,7 +33,8 @@ router.post('/', checkAuth, multer({ storage: storage }).single('image'), (req, 
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save()
     .then((createdPost) => {
@@ -50,17 +52,23 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req
   let imagePath = req.body.imagePath;
   if (req.file) {
     const url = req.protocol + '://' + req.get('host');
-    imagePath = url + '/images/' + req.file.filename
+    imagePath = url + '/images/' + req.file.filename;
   }
   const post = new Post({
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({ _id: req.params.id }, post)
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
     .then((result) => {
-      res.status(200).json({ message: 'Update successful!' });
+      if(result.modifiedCount > 0) {
+        res.status(200).json({ message: 'Update successful!' });
+      }
+      else {
+        res.status(401).json({ message: 'Not authorized!' })
+      }
     });
 });
 
@@ -69,7 +77,7 @@ router.get('/', (req, res, next) => {
   const currentPage = +req.query.page;
   const postQuery = Post.find();
   let fetchedPosts;
-  if(pageSize && currentPage) {
+  if (pageSize && currentPage) {
     postQuery
       .skip(pageSize * (currentPage - 1))
       .limit(pageSize);
@@ -98,16 +106,21 @@ router.get('/:id', (req, res, next) => {
         res.status(200).json(post);
       }
       else {
-        res.status(400).json({ message: 'Post not found!' });
+        res.status(401).json({ message: 'Post not found!' });
       }
     });
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
   const id = req.params.id;
-  Post.deleteOne({ _id: id })
-    .then(() => {
-      res.status(200).json({ message: 'Post deleted' });
+  Post.deleteOne({ _id: id, creator: req.userData.userId })
+    .then((result) => {
+      if(result.deletedCount > 0) {
+        res.status(200).json({ message: 'Post deleted' });
+      }
+      else {
+        res.status(401).json({ message: 'Post not found!' });
+      }
     });
 });
